Add card classes, implement basic game rules
Borrowed Card, Deck, and Hand classes from Old Maid example
introduced in MIT OCW Python notes. The game begins with the
center card filled in. Game board now shows the active card
of the current player along the bottom of the screen, which
can be placed onto any square in the grid not in the center.

To be implemented:
 - Fix cards being added to the same grid square repeatedly
 - Replace facedown cards on the grid with green rectangles
 - Make sure green rectangles replace each players' hands
   when all cards have been played