Initial PvC commit -- Looks sloppy
Implemented:
* Challenge Mode allows play with a computer challenger.
* Computer calculates which spaces allow for maximum scoring
differential and randomly places card in one of these spaces.
* Changed flush scoring so that the computer discounts flushes in the
early stages of the game.
* Buttons are removed after computer chooses blank spaces, preventing
double-placement. Grid buttons are all now in a list, making it easy
to operate on them.

Still to be implemented:
* Disable grid buttons during animations.
* Display more information according to the mode, e.g. scoring goal,
high scores.
* Clean up how the way a new round loads up. Grid guides need to be in
the right places, the right player needs to be active, etc.
* Refactor code so that there are less fiddly little exceptions. Make
things streamlined.
* Add difficulty levels for challenge mode.
* Allow players to choose names.
* Add pause
* Add quit to menu
* Add settings menu
* Improve visuals. Add fancy title, make buttons pretty, add pretty
card placement and card flipping animations.